Status / Monitoring - Quick Links II - Genesis
Quick links provides convenient one click direct access to graphs without the need to reload the entire page or find the desired graph in the saved views list.

1) More convenient and quicker than expanding the saved views drop down and searching through the list for the desired graph.

2) Refreshes only the graph.  This is faster and more efficient than reloading the entire page by selecting a saved view.

This can be useful for quick convenient access to temporary graphs being used for issue investigation and troubleshooting as well as frequently accessed permanent graphs.